Edit CartController

+++ FinalProjectDominos/src/main/java/ittalents/dominos/model/DTOs/ProductWithoutImageDTO.java
@@ -1,12 +1,13 @@
 package ittalents.dominos.model.DTOs;
 
-import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -20,7 +21,7 @@ public class ProductWithoutImageDTO {
 
     private String name;
     @DecimalMin("0.0")
-    private double price;
+    private BigDecimal price;
 
     private int categoryId;
